+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add `transaction update-supply-plt` command to mint or burn protocol level tokens by the governance account.
+
 ## 9.0.0 - 2025-05-09
 
 - Add `raw GetTokenInfo` command to display the info of a protocol level token.